+++ lib/arel/extensions.rb
@@ -46,6 +46,7 @@ require 'arel/extensions/contains'
 require 'arel/extensions/contained_by'
 require 'arel/extensions/overlap'
 require 'arel/extensions/select_statement'
+require 'arel/extensions/insert_statement'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
