+++ lib/arel/extensions/nodes.rb
@@ -5,14 +5,6 @@
 module Arel
   module Nodes
     # https://www.postgresql.org/docs/9.5/sql-insert.html
-    Arel::Nodes::InsertStatement.class_eval do
-      attr_accessor :with
-      attr_accessor :on_conflict
-      attr_accessor :override
-      attr_accessor :returning
-    end
-
-    # https://www.postgresql.org/docs/9.5/sql-insert.html
     class DefaultValues < Arel::Nodes::Node
     end
