+++ lib/arel/extensions/visitors.rb
@@ -52,54 +52,6 @@ module Arel
         collector << 'DEFAULT VALUES'
       end
 
-      # rubocop:disable Metrics/CyclomaticComplexity
-      # rubocop:disable Metrics/AbcSize
-      # rubocop:disable Metrics/PerceivedComplexity
-      def visit_Arel_Nodes_InsertStatement(o, collector)
-        if o.with
-          collector = visit o.with, collector
-          collector << SPACE
-        end
-
-        collector << 'INSERT INTO '
-        collector = visit o.relation, collector
-        if o.columns.any?
-          collector << " (#{o.columns.map do |x|
-            quote_column_name x.name
-          end.join ', '})"
-        end
-
-        case o.override
-        when 0
-          collector << ''
-        when 1
-          collector << ' OVERRIDING USER VALUE'
-        when 2
-          collector << ' OVERRIDING SYSTEM VALUE'
-        else
-          raise "Unknown override `#{o.override}`"
-        end
-
-        collector = if o.values
-                      maybe_visit o.values, collector
-                    elsif o.select
-                      maybe_visit o.select, collector
-                    else
-                      collector
-                    end
-
-        unless o.returning.empty?
-          collector << ' RETURNING '
-          collector = inject_join o.returning, collector, ', '
-        end
-
-        visit(o.on_conflict, collector) if o.on_conflict
-        collector
-      end
-      # rubocop:enable Metrics/CyclomaticComplexity
-      # rubocop:enable Metrics/AbcSize
-      # rubocop:enable Metrics/PerceivedComplexity
-
       # rubocop:disable Metrics/AbcSize
       def visit_Arel_Nodes_Conflict(o, collector)
         collector << ' ON CONFLICT '
